Fix all scrolled panels code
