+++ meerk40t/gui/choicepropertypanel.py
@@ -1,9 +1,8 @@
 import wx
-from wx.lib.scrolledpanel import ScrolledPanel
 
 from meerk40t.core.units import Angle, Length
 from meerk40t.gui.laserrender import swizzlecolor
-from meerk40t.gui.wxutils import TextCtrl
+from meerk40t.gui.wxutils import TextCtrl, ScrolledPanel
 from meerk40t.kernel import Context
 from meerk40t.svgelements import Color
