+++ meerk40t/gui/propertypanels/operationpropertymain.py
@@ -1,5 +1,5 @@
 import wx
-from wx.lib.scrolledpanel import ScrolledPanel
+from meerk40t.gui.wxutils import ScrolledPanel
 
 from meerk40t.kernel import signal_listener
