+++ meerk40t/lihuiyu/gui/lhycontrollergui.py
@@ -4,7 +4,7 @@
 #
 
 import wx
-from wx.lib.scrolledpanel import ScrolledPanel
+from meerk40t.gui.wxutils import ScrolledPanel
 
 from meerk40t.gui.icons import (
     icons8_connected_50,
